+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/processing/GregtechMetaTileEntity_IndustrialAlloySmelter.java
@@ -247,7 +247,7 @@ public class GregtechMetaTileEntity_IndustrialAlloySmelter extends GregtechMeta_
 	
 	@Override
 	public boolean checkRecipeGeneric(ItemStack[] aItemInputs, FluidStack[] aFluidInputs, int aMaxParallelRecipes,
-			int aEUPercent, int aSpeedBonusPercent, int aOutputChanceRoll) {
+			long aEUPercent, int aSpeedBonusPercent, int aOutputChanceRoll) {
 		// Based on the Processing Array. A bit overkill, but very flexible.
 
 		// Reset outputs and progress stats
